requirements.txt
asgiref==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
brotli==1.0.9 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4.0"
cfgv==3.3.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.10" and python_version < "4"
cloudinary==1.32.0 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0"
distlib==0.3.6 ; python_version >= "3.10" and python_version < "4.0"
dj-database-url==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
django-cloudinary-storage==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==3.14.0 ; python_version >= "3.10" and python_version < "4.0"
django-rest-framework==0.1.0 ; python_version >= "3.10" and python_version < "4.0"
django==4.1.7 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.9.0 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
identify==2.5.18 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
nodeenv==1.7.0 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.4.0 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
pre-commit==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2==2.9.5 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
setuptools==67.4.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
uql @ git+https://github.com/rubbieKelvin/UntitledQL.git ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.10" and python_version < "4.0"
virtualenv==20.19.0 ; python_version >= "3.10" and python_version < "4.0"
whitenoise[brotli]==6.4.0 ; python_version >= "3.10" and python_version < "4.0"
